Rearrange after merge
