+++ src/Auditable.php
@@ -61,6 +61,13 @@ trait Auditable
     public $preloadedResolverData = [];
 
     /**
+     * Auditable attributes excluded from the Audit.
+     *
+     * @var array
+     */
+    protected $excludedAttributes = [];
+
+    /**
      * Auditable boot logic.
      *
      * @return void
@@ -102,7 +109,7 @@ trait Auditable
     /**
      * Modify attribute value.
      *
-     * @param  mixed  $value
+     * @param mixed $value
      * @return mixed
      *
      * @throws AuditingException
@@ -111,7 +118,7 @@ trait Auditable
     {
         $attributeModifiers = $this->getAttributeModifiers();
 
-        if (! array_key_exists($attribute, $attributeModifiers)) {
+        if (!array_key_exists($attribute, $attributeModifiers)) {
             return $value;
         }
 
@@ -128,152 +135,13 @@ trait Auditable
         throw new AuditingException(sprintf('Invalid AttributeModifier implementation: %s', $attributeModifier));
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function toAudit(): array
-    {
-        if (!$this->readyForAuditing()) {
-            throw new AuditingException('A valid audit event has not been set');
-        }
 
-        $attributeGetter = $this->resolveAttributeGetter($this->auditEvent);
 
-        if (!method_exists($this, $attributeGetter)) {
-            throw new AuditingException(sprintf(
-                'Unable to handle "%s" event, %s() method missing',
-                $this->auditEvent,
-                $attributeGetter
-            ));
-        }
-
-        $this->resolveAuditExclusions();
-
-        list($old, $new) = $this->$attributeGetter();
-
-        if ($this->getAttributeModifiers() && !$this->isCustomEvent) {
-            foreach ($old as $attribute => $value) {
-                $old[$attribute] = $this->modifyAttributeValue($attribute, $value);
-            }
-
-            foreach ($new as $attribute => $value) {
-                $new[$attribute] = $this->modifyAttributeValue($attribute, $value);
-            }
-        }
-
-        $morphPrefix = Config::get('audit.user.morph_prefix', 'user');
-
-        $tags = implode(',', $this->generateTags());
-
-        $user = $this->resolveUser();
-
-        return $this->transformAudit(array_merge([
-            'old_values'           => $old,
-            'new_values'           => $new,
-            'event'                => $this->auditEvent,
-            'auditable_id'         => $this->getKey(),
-            'auditable_type'       => $this->getMorphClass(),
-            $morphPrefix . '_id'   => $user ? $user->getAuthIdentifier() : null,
-            $morphPrefix . '_type' => $user ? $user->getMorphClass() : null,
-            'tags'                 => empty($tags) ? null : $tags,
-        ], $this->runResolvers()));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function transformAudit(array $data): array
-    {
-        return $data;
-    }
-
-    /**
-     * Resolve the User.
-     *
-     * @return mixed|null
-     * @throws AuditingException
-     *
-     */
-    protected function resolveUser()
-    {
-        $userResolver = Config::get('audit.user.resolver');
-
-        if (is_null($userResolver) && Config::has('audit.resolver') && !Config::has('audit.user.resolver')) {
-            trigger_error(
-                'The config file audit.php is not updated to the new version 13.0. Please see https://laravel-auditing.com/guide/upgrading.html',
-                E_USER_DEPRECATED
-            );
-            $userResolver = Config::get('audit.resolver.user');
-        }
-
-        if (is_subclass_of($userResolver, \OwenIt\Auditing\Contracts\UserResolver::class)) {
-            return call_user_func([$userResolver, 'resolve'], $this);
-        }
-
-        throw new AuditingException('Invalid UserResolver implementation');
-    }
-
-    protected function runResolvers(): array
-    {
-        $resolved = [];
-        $resolvers = Config::get('audit.resolvers', []);
-        if (empty($resolvers) && Config::has('audit.resolver')) {
-            trigger_error(
-                'The config file audit.php is not updated to the new version 13.0. Please see https://laravel-auditing.com/guide/upgrading.html',
-                E_USER_DEPRECATED
-            );
-            $resolvers = Config::get('audit.resolver', []);
-        }
-
-        foreach ($resolvers as $name => $implementation) {
-            if (empty($implementation)) {
-                continue;
-            }
-
-            if (!is_subclass_of($implementation, Resolver::class)) {
-                throw new AuditingException('Invalid Resolver implementation for: ' . $name);
-            }
-            $resolved[$name] = call_user_func([$implementation, 'resolve'], $this);
-        }
-        return $resolved;
-    }
-
-    public function preloadResolverData()
-    {
-        $this->preloadedResolverData = $this->runResolvers();
-
-        if (!empty ($this->resolveUser())) {
-            $this->preloadedResolverData['user'] = $this->resolveUser();
-        }
-
-        return $this;
-    }
-
-    /**
-     * Determine if an attribute is eligible for auditing.
-     *
-     * @param string $attribute
-     *
-     * @return bool
-     */
-    protected function isAttributeAuditable(string $attribute): bool
-    {
-        // The attribute should not be audited
-        if (in_array($attribute, $this->excludedAttributes, true)) {
-            return false;
-        }
-
-        // The attribute is auditable when explicitly
-        // listed or when the include array is empty
-        $include = $this->getAuditInclude();
-
-        return empty($include) || in_array($attribute, $include, true);
-    }
 
     /**
      * Determine whether an event is auditable.
      *
-     * @param  string  $event
+     * @param string $event
      */
     protected function isEventAuditable($event): bool
     {
@@ -383,178 +251,7 @@ trait Auditable
         return [];
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function transitionTo(Contracts\Audit $audit, bool $old = false): Contracts\Auditable
-    {
-        // The Audit must be for an Auditable model of this type
-        if ($this->getMorphClass() !== $audit->auditable_type) {
-            throw new AuditableTransitionException(sprintf(
-                'Expected Auditable type %s, got %s instead',
-                $this->getMorphClass(),
-                $audit->auditable_type
-            ));
-        }
 
-        // The Audit must be for this specific Auditable model
-        if ($this->getKey() !== $audit->auditable_id) {
-            throw new AuditableTransitionException(sprintf(
-                'Expected Auditable id (%s)%s, got (%s)%s instead',
-                gettype($this->getKey()),
-                $this->getKey(),
-                gettype($audit->auditable_id),
-                $audit->auditable_id
-            ));
-        }
 
-        // Redacted data should not be used when transitioning states
-        foreach ($this->getAttributeModifiers() as $attribute => $modifier) {
-            if (is_subclass_of($modifier, AttributeRedactor::class)) {
-                throw new AuditableTransitionException('Cannot transition states when an AttributeRedactor is set');
-            }
-        }
 
-        // The attribute compatibility between the Audit and the Auditable model must be met
-        $modified = $audit->getModified();
-
-        if ($incompatibilities = array_diff_key($modified, $this->getAttributes())) {
-            throw new AuditableTransitionException(sprintf(
-                'Incompatibility between [%s:%s] and [%s:%s]',
-                $this->getMorphClass(),
-                $this->getKey(),
-                get_class($audit),
-                $audit->getKey()
-            ), array_keys($incompatibilities));
-        }
-
-        $key = $old ? 'old' : 'new';
-
-        foreach ($modified as $attribute => $value) {
-            if (array_key_exists($key, $value)) {
-                $this->setAttribute($attribute, $value[$key]);
-            }
-        }
-
-        return $this;
-    }
-
-    /*
-    |--------------------------------------------------------------------------
-    | Pivot help methods
-    |--------------------------------------------------------------------------
-    |
-    | Methods for auditing pivot actions
-    |
-    */
-
-    /**
-     * @param string $relationName
-     * @param mixed $id
-     * @param array $attributes
-     * @param bool $touch
-     * @param array $columns
-     * @return void
-     * @throws AuditingException
-     */
-    public function auditAttach(string $relationName, $id, array $attributes = [], $touch = true, $columns = ['*'])
-    {
-        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'attach')) {
-            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method attach');
-        }
-
-        $old = $this->{$relationName}()->get($columns);
-        $this->{$relationName}()->attach($id, $attributes, $touch);
-        $new = $this->{$relationName}()->get($columns);
-        $this->dispatchRelationAuditEvent($relationName, 'attach', $old, $new);
-    }
-
-    /**
-     * @param string $relationName
-     * @param mixed $ids
-     * @param bool $touch
-     * @param array $columns
-     * @return int
-     * @throws AuditingException
-     */
-    public function auditDetach(string $relationName, $ids = null, $touch = true, $columns = ['*'])
-    {
-        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'detach')) {
-            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method detach');
-        }
-
-        $old = $this->{$relationName}()->get($columns);
-        $results = $this->{$relationName}()->detach($ids, $touch);
-        $new = $this->{$relationName}()->get($columns);
-        $this->dispatchRelationAuditEvent($relationName, 'detach', $old, $new);
-
-        return empty($results) ? 0 : $results;
-    }
-
-    /**
-     * @param $relationName
-     * @param \Illuminate\Support\Collection|\Illuminate\Database\Eloquent\Model|array $ids
-     * @param bool $detaching
-     * @param array $columns
-     * @return array
-     * @throws AuditingException
-     */
-    public function auditSync($relationName, $ids, $detaching = true, $columns = ['*'])
-    {
-        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'sync')) {
-            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method sync');
-        }
-
-        $old = $this->{$relationName}()->get($columns);
-        $changes = $this->{$relationName}()->sync($ids, $detaching);
-        if (collect($changes)->flatten()->isEmpty()) {
-            $old = $new = collect([]);
-        } else {
-            $new = $this->{$relationName}()->get($columns);
-        }
-        $this->dispatchRelationAuditEvent($relationName, 'sync', $old, $new);
-
-        return $changes;
-    }
-
-    /**
-     * @param string $relationName
-     * @param \Illuminate\Support\Collection|\Illuminate\Database\Eloquent\Model|array $ids
-     * @param array $columns
-     * @return array
-     * @throws AuditingException
-     */
-    public function auditSyncWithoutDetaching(string $relationName, $ids, $columns = ['*'])
-    {
-        if (!method_exists($this, $relationName) || !method_exists($this->{$relationName}(), 'syncWithoutDetaching')) {
-            throw new AuditingException('Relationship ' . $relationName . ' was not found or does not support method syncWithoutDetaching');
-        }
-
-        return $this->auditSync($relationName, $ids, false, $columns);
-    }
-
-    /**
-     * @param string $relationName
-     * @param string $event
-     * @param \Illuminate\Support\Collection $old
-     * @param \Illuminate\Support\Collection $new
-     * @return void
-     */
-    private function dispatchRelationAuditEvent($relationName, $event, $old, $new)
-    {
-        $this->auditCustomOld[$relationName] = $old->diff($new)->toArray();
-        $this->auditCustomNew[$relationName] = $new->diff($old)->toArray();
-
-        if (
-            empty($this->auditCustomOld[$relationName]) &&
-            empty($this->auditCustomNew[$relationName])
-        ) {
-            $this->auditCustomOld = $this->auditCustomNew = [];
-        }
-
-        $this->auditEvent = $event;
-        $this->isCustomEvent = true;
-        Event::dispatch(AuditCustom::class, [$this]);
-        $this->isCustomEvent = false;
-    }
 }
